Trying to track down the micronas loss-of-sync glitch. The plot thickens....
Here is a standalone tool I just made which scans the last 2KB of a file
searching for the last valid and complete mp3 frame. Most of my mp3s were
ripped 2+ years ago, some using lame3.22 and some using audiocatalyst.

All of the lame3.22 files I've looked at (without id3 tags) appear to be quite
horribly broken at the end of the file. Some end with a valid frame followed
by a useless 0, others end with a valid frame followed by another frame header
and a truncated frame. So strange. The Audiocatalyst files look okay.

This is turning out to be a bigger PIA than it seemed at first glance, but
at least we know that it's broken files that cause this, and it's probably
fixable for most files.

My only concern is that the files which end in a truncated frame may
by quite hard to repair if there's anything in the bit reservoir at that point.


git-svn-id: file:///home/awy/mirror/slim/trunk/tools@106 62299810-d8cb-41fd-93b7-d32162e5a4a4