Protocol tweaks on client and server.  You'll need to update firmware.
Fixed some problems that could cause a crash with broken plugins.

Accepted patch for slimp3.pl

Simplistic work to make sure that we give time to both slimp3 and squeezebox for streaming.

Patch to fix sorting of albums with mixed artists.

Half of the work to display wireless strength in Info plugin.  Need to supress on slimp3 and squeezebox w/ ethernet.


git-svn-id: file:///home/awy/mirror/slim/trunk/tools@193 62299810-d8cb-41fd-93b7-d32162e5a4a4